requirements.txt
neo4j==5.14.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.24.3
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1